requirements.txt
torch>=2.4.0
torchvision>=0.19
timm>=1.0.7
pyyaml>=6.0
